--- a/graph/community/louvain_directed.go
+++ b/graph/community/louvain_directed.go
@@ -6,7 +6,7 @@ package community
 
 import (
 	"math"
-	"sort"
+	"slices"
 
 	"golang.org/x/exp/rand"
 
@@ -617,7 +617,7 @@ func (l *directedLocalMover) deltaQ(n graph.Node) (deltaQ float64, dst int, src 
 	for i := range connected {
 		candidates = append(candidates, i)
 	}
-	sort.Ints(candidates)
+	slices.Sort(candidates)
 
 	// Calculate the highest modularity gain
 	// from moving into another community and