--- a/graph/community/louvain_undirected_multiplex_test.go
+++ b/graph/community/louvain_undirected_multiplex_test.go
@@ -9,7 +9,6 @@ import (
 	"math"
 	"reflect"
 	"slices"
-	"sort"
 	"testing"
 
 	"golang.org/x/exp/rand"
@@ -578,7 +577,7 @@ func TestLouvainMultiplex(t *testing.T) {
 			}
 
 			// Recovery of Q values is reversed.
-			if slices.Reverse(qs); !sort.Float64sAreSorted(qs) {
+			if slices.Reverse(qs); !slices.IsSorted(qs) {
 				t.Errorf("Q values not monotonically increasing: %.5v", qs)
 			}
 		}