@jbduncan 2024-03-16
> This needs to pick up the changes in master.

Done. I'll pick them up again if https://github.com/gonum/gonum/pull/1953 is merged in.