@jbduncan 2024-04-14
> Please reorder the commits so that the benchmarks are introduced first and then show the perf impacts of the change.

Okay, commits re-ordered!

I intend to rebase and run the benchmarks as soon as prerequisite PR #1953 is merged in.