@jbduncan 2024-05-24
> What are the benchmarking outcomes from the non-`coloring.Set` changes?

Thanks for checking! I'm not finished with my benchmarks yet. I ran them last night, but I made a mistake where I ran them on this branch and `master` whereas I should have run them on both commits in this branch instead. Stay tuned.