@jbduncan 2024-05-27
Here are the benchmarks. I'm not sure what to make of them. WDYT?

<details>
  <summary>Benchmarks</summary>
   
  ```
  goos: darwin
goarch: arm64
pkg: gonum.org/v1/gonum/graph/coloring
                                                       │  bench.old   │               bench.new               │
                                                       │    sec/op    │    sec/op      vs base                │
Coloring/empty/Dsatur-8                                  3.196n ±  0%    3.191n ±  0%        ~ (p=0.164 n=10)
Coloring/empty/DsaturExact-8                             3.191n ±  1%    3.191n ±  0%        ~ (p=0.489 n=10)
Coloring/empty/Randomized-8                              3.192n ±  0%    3.191n ±  0%        ~ (p=0.722 n=10)
Coloring/empty/RecursiveLargestFirst-8                   3.191n ±  0%    3.191n ±  0%        ~ (p=0.180 n=10)
Coloring/empty/SanSegundo-8                              3.190n ±  0%    3.191n ±  0%        ~ (p=0.329 n=10)
Coloring/empty/WelshPowell-8                             3.191n ±  1%    3.191n ±  0%        ~ (p=0.512 n=10)
Coloring/singleton/Dsatur-8                              414.6n ±  0%    416.8n ±  0%   +0.54% (p=0.000 n=10)
Coloring/singleton/DsaturExact-8                         962.6n ±  0%    965.2n ±  0%   +0.26% (p=0.037 n=10)
Coloring/singleton/Randomized-8                          236.7n ±  1%    238.5n ±  0%   +0.74% (p=0.022 n=10)
Coloring/singleton/RecursiveLargestFirst-8               298.7n ±  0%    298.9n ±  0%        ~ (p=0.156 n=10)
Coloring/singleton/SanSegundo-8                          415.3n ±  0%    417.1n ±  0%   +0.43% (p=0.001 n=10)
Coloring/singleton/WelshPowell-8                         289.9n ±  0%    291.6n ±  0%   +0.57% (p=0.001 n=10)
Coloring/kite/Dsatur-8                                   3.332µ ±  2%    3.384µ ±  0%   +1.58% (p=0.022 n=10)
Coloring/kite/DsaturExact-8                              17.99µ ±  0%    18.26µ ±  1%   +1.46% (p=0.000 n=10)
Coloring/kite/Randomized-8                               1.177µ ±  0%    1.197µ ±  0%   +1.70% (p=0.000 n=10)
Coloring/kite/RecursiveLargestFirst-8                    2.267µ ±  0%    2.269µ ±  0%   +0.09% (p=0.015 n=10)
Coloring/kite/SanSegundo-8                               4.043µ ±  0%    4.106µ ±  0%   +1.55% (p=0.000 n=10)
Coloring/kite/WelshPowell-8                              1.410µ ±  0%    1.435µ ±  0%   +1.77% (p=0.001 n=10)
Coloring/triangle+1/Dsatur-8                             2.106µ ±  0%    2.147µ ±  1%   +1.95% (p=0.000 n=10)
Coloring/triangle+1/DsaturExact-8                        10.14µ ±  0%    10.26µ ±  0%   +1.24% (p=0.000 n=10)
Coloring/triangle+1/Randomized-8                         820.3n ±  0%    839.7n ±  0%   +2.37% (p=0.000 n=10)
Coloring/triangle+1/RecursiveLargestFirst-8              1.453µ ±  1%    1.457µ ±  0%   +0.24% (p=0.048 n=10)
Coloring/triangle+1/SanSegundo-8                         2.452µ ±  0%    2.491µ ±  0%   +1.61% (p=0.000 n=10)
Coloring/triangle+1/WelshPowell-8                        983.6n ±  0%   1002.0n ±  2%   +1.88% (p=0.000 n=10)
Coloring/bipartite_halves/Dsatur-8                       8.054µ ±  0%    8.176µ ±  0%   +1.51% (p=0.000 n=10)
Coloring/bipartite_halves/DsaturExact-8                  33.32µ ±  1%    33.68µ ±  0%   +1.07% (p=0.001 n=10)
Coloring/bipartite_halves/Randomized-8                   2.196µ ±  0%    2.218µ ±  0%   +1.00% (p=0.000 n=10)
Coloring/bipartite_halves/RecursiveLargestFirst-8        4.027µ ±  0%    4.010µ ±  0%   -0.42% (p=0.030 n=10)
Coloring/bipartite_halves/SanSegundo-8                   15.25µ ±  0%    15.43µ ±  0%   +1.18% (p=0.000 n=10)
Coloring/bipartite_halves/WelshPowell-8                  2.401µ ±  0%    2.447µ ±  0%   +1.90% (p=0.000 n=10)
Coloring/bipartite_alternating/Dsatur-8                  7.924µ ±  2%    8.061µ ±  0%   +1.73% (p=0.022 n=10)
Coloring/bipartite_alternating/DsaturExact-8             33.12µ ±  0%    33.53µ ±  0%   +1.22% (p=0.002 n=10)
Coloring/bipartite_alternating/Randomized-8              2.178µ ±  0%    2.204µ ±  1%   +1.22% (p=0.000 n=10)
Coloring/bipartite_alternating/RecursiveLargestFirst-8   3.986µ ±  0%    3.975µ ±  0%   -0.30% (p=0.000 n=10)
Coloring/bipartite_alternating/SanSegundo-8              15.37µ ±  1%    15.50µ ±  0%   +0.82% (p=0.003 n=10)
Coloring/bipartite_alternating/WelshPowell-8             2.534µ ±  1%    2.536µ ±  0%        ~ (p=0.322 n=10)
Coloring/3/4_bipartite/Dsatur-8                          6.468µ ±  0%    6.548µ ±  0%   +1.23% (p=0.000 n=10)
Coloring/3/4_bipartite/DsaturExact-8                     30.60µ ±  0%    30.91µ ±  1%   +1.02% (p=0.000 n=10)
Coloring/3/4_bipartite/Randomized-8                      1.869µ ±  0%    1.896µ ±  0%   +1.47% (p=0.000 n=10)
Coloring/3/4_bipartite/RecursiveLargestFirst-8           3.566µ ±  1%    3.549µ ±  0%   -0.46% (p=0.000 n=10)
Coloring/3/4_bipartite/SanSegundo-8                      13.46µ ±  0%    13.56µ ±  0%   +0.80% (p=0.000 n=10)
Coloring/3/4_bipartite/WelshPowell-8                     2.250µ ±  0%    2.279µ ±  0%   +1.29% (p=0.001 n=10)
Coloring/cubical/Dsatur-8                                8.074µ ±  0%    8.184µ ±  0%   +1.37% (p=0.000 n=10)
Coloring/cubical/DsaturExact-8                           33.55µ ±  0%    33.89µ ±  0%   +1.02% (p=0.000 n=10)
Coloring/cubical/Randomized-8                            2.211µ ±  2%    2.232µ ±  1%   +0.95% (p=0.017 n=10)
Coloring/cubical/RecursiveLargestFirst-8                 4.097µ ±  0%    4.089µ ±  0%   -0.18% (p=0.022 n=10)
Coloring/cubical/SanSegundo-8                            15.43µ ±  0%    15.56µ ±  0%   +0.83% (p=0.000 n=10)
Coloring/cubical/WelshPowell-8                           2.522µ ±  0%    2.562µ ±  0%   +1.61% (p=0.000 n=10)
Coloring/HoG33106/Dsatur-8                               19.15µ ±  0%    19.37µ ±  0%   +1.11% (p=0.000 n=10)
Coloring/HoG33106/DsaturExact-8                          74.02µ ±  0%    74.72µ ±  0%   +0.94% (p=0.002 n=10)
Coloring/HoG33106/Randomized-8                           4.056µ ±  0%    4.082µ ±  1%   +0.65% (p=0.000 n=10)
Coloring/HoG33106/RecursiveLargestFirst-8                8.855µ ±  0%    8.848µ ±  0%        ~ (p=0.196 n=10)
Coloring/HoG33106/SanSegundo-8                           69.19µ ±  1%    69.78µ ±  0%   +0.85% (p=0.003 n=10)
Coloring/HoG33106/WelshPowell-8                          5.007µ ±  0%    5.056µ ±  0%   +1.00% (p=0.000 n=10)
Coloring/HoG41237/Dsatur-8                               206.0m ±  0%    212.4m ±  1%   +3.09% (p=0.000 n=10)
Coloring/HoG41237/DsaturExact-8                          225.3m ±  0%    232.4m ±  0%   +3.15% (p=0.000 n=10)
Coloring/HoG41237/Randomized-8                           2.411m ±  2%    2.405m ±  0%   -0.22% (p=0.007 n=10)
Coloring/HoG41237/RecursiveLargestFirst-8                8.755m ±  0%    8.810m ±  1%   +0.63% (p=0.001 n=10)
Coloring/HoG41237/SanSegundo-8                           495.9m ±  1%    501.5m ±  0%   +1.12% (p=0.003 n=10)
Coloring/HoG41237/WelshPowell-8                          3.581m ±  1%    3.606m ±  1%   +0.69% (p=0.009 n=10)
Coloring/exams/Dsatur-8                                  4.686µ ±  0%    4.742µ ±  0%   +1.21% (p=0.000 n=10)
Coloring/exams/DsaturExact-8                             26.82µ ±  0%    27.17µ ±  1%   +1.29% (p=0.000 n=10)
Coloring/exams/Randomized-8                              1.524µ ±  0%    1.545µ ±  0%   +1.34% (p=0.000 n=10)
Coloring/exams/RecursiveLargestFirst-8                   3.108µ ±  2%    3.079µ ±  0%   -0.95% (p=0.000 n=10)
Coloring/exams/SanSegundo-8                              7.514µ ±  0%    7.582µ ±  0%   +0.91% (p=0.000 n=10)
Coloring/exams/WelshPowell-8                             1.847µ ±  0%    1.870µ ±  0%   +1.25% (p=0.000 n=10)
Coloring/exam_scheduler_1/Dsatur-8                       18.09µ ±  0%    18.32µ ±  1%   +1.25% (p=0.000 n=10)
Coloring/exam_scheduler_1/DsaturExact-8                  100.4µ ±  0%    101.5µ ±  0%   +1.08% (p=0.000 n=10)
Coloring/exam_scheduler_1/Randomized-8                   4.322µ ±  2%    4.353µ ±  0%   +0.71% (p=0.023 n=10)
Coloring/exam_scheduler_1/RecursiveLargestFirst-8        11.49µ ±  0%    11.46µ ±  0%   -0.27% (p=0.006 n=10)
Coloring/exam_scheduler_1/SanSegundo-8                   27.86µ ±  0%    28.10µ ±  0%   +0.89% (p=0.000 n=10)
Coloring/exam_scheduler_1/WelshPowell-8                  4.759µ ±  0%    4.834µ ±  0%   +1.58% (p=0.000 n=10)
Coloring/exam_scheduler_2/Dsatur-8                       8.432µ ±  0%    8.552µ ±  0%   +1.42% (p=0.000 n=10)
Coloring/exam_scheduler_2/DsaturExact-8                  48.77µ ±  0%    49.02µ ±  0%   +0.53% (p=0.001 n=10)
Coloring/exam_scheduler_2/Randomized-8                   2.371µ ±  0%    2.396µ ±  0%   +1.03% (p=0.000 n=10)
Coloring/exam_scheduler_2/RecursiveLargestFirst-8        5.243µ ±  0%    5.233µ ±  1%   -0.19% (p=0.022 n=10)
Coloring/exam_scheduler_2/SanSegundo-8                   12.89µ ±  0%    13.08µ ±  0%   +1.44% (p=0.000 n=10)
Coloring/exam_scheduler_2/WelshPowell-8                  2.769µ ±  4%    2.809µ ±  0%   +1.44% (p=0.022 n=10)
Coloring/Brélaz/Dsatur-8                                 12.23µ ±  0%    12.36µ ±  0%   +1.04% (p=0.000 n=10)
Coloring/Brélaz/DsaturExact-8                            60.25µ ±  0%    60.79µ ±  1%   +0.90% (p=0.000 n=10)
Coloring/Brélaz/Randomized-8                             3.116µ ±  0%    3.142µ ±  0%   +0.83% (p=0.000 n=10)
Coloring/Brélaz/RecursiveLargestFirst-8                  7.655µ ±  1%    7.645µ ±  0%   -0.14% (p=0.018 n=10)
Coloring/Brélaz/SanSegundo-8                             18.62µ ±  0%    18.75µ ±  0%   +0.68% (p=0.000 n=10)
Coloring/Brélaz/WelshPowell-8                            3.743µ ±  0%    3.784µ ±  0%   +1.11% (p=0.000 n=10)
Coloring/Lima_Carmo/Dsatur-8                             7.617µ ±  0%    7.723µ ±  1%   +1.39% (p=0.000 n=10)
Coloring/Lima_Carmo/DsaturExact-8                        33.09µ ±  0%    33.41µ ±  0%   +0.97% (p=0.000 n=10)
Coloring/Lima_Carmo/Randomized-8                         2.107µ ±  1%    2.124µ ±  0%   +0.78% (p=0.019 n=10)
Coloring/Lima_Carmo/RecursiveLargestFirst-8              4.242µ ±  0%    4.216µ ±  0%   -0.62% (p=0.000 n=10)
Coloring/Lima_Carmo/SanSegundo-8                         11.35µ ±  0%    11.48µ ±  0%   +1.13% (p=0.000 n=10)
Coloring/Lima_Carmo/WelshPowell-8                        2.502µ ±  0%    2.532µ ±  2%   +1.20% (p=0.000 n=10)
Coloring/San_Segundo/Dsatur-8                            12.92µ ±  0%    13.03µ ±  0%   +0.88% (p=0.000 n=10)
Coloring/San_Segundo/DsaturExact-8                       121.3µ ±  2%    121.8µ ±  0%   +0.37% (p=0.029 n=10)
Coloring/San_Segundo/Randomized-8                        3.484µ ±  0%    3.498µ ±  0%   +0.40% (p=0.000 n=10)
Coloring/San_Segundo/RecursiveLargestFirst-8             9.289µ ±  0%    9.242µ ±  0%   -0.51% (p=0.001 n=10)
Coloring/San_Segundo/SanSegundo-8                        22.82µ ±  0%    22.94µ ±  0%   +0.50% (p=0.019 n=10)
Coloring/San_Segundo/WelshPowell-8                       3.635µ ±  2%    3.684µ ±  5%   +1.33% (p=0.006 n=10)
Coloring/tetrahedron/Dsatur-8                            2.415µ ±  0%    2.466µ ±  3%   +2.13% (p=0.000 n=10)
Coloring/tetrahedron/DsaturExact-8                       13.75µ ±  0%    13.91µ ±  0%   +1.15% (p=0.000 n=10)
Coloring/tetrahedron/Randomized-8                        941.4n ±  0%    962.0n ±  1%   +2.19% (p=0.000 n=10)
Coloring/tetrahedron/RecursiveLargestFirst-8             1.890µ ±  1%    1.883µ ±  0%   -0.40% (p=0.000 n=10)
Coloring/tetrahedron/SanSegundo-8                        3.370µ ±  0%    3.408µ ±  0%   +1.13% (p=0.001 n=10)
Coloring/tetrahedron/WelshPowell-8                       1.078µ ±  0%    1.099µ ±  0%   +2.00% (p=0.000 n=10)
Coloring/triangle/Dsatur-8                               1.498µ ±  0%    1.529µ ±  2%   +2.04% (p=0.000 n=10)
Coloring/triangle/DsaturExact-8                          7.365µ ±  0%    7.390µ ±  0%   +0.33% (p=0.000 n=10)
Coloring/triangle/Randomized-8                           647.1n ±  0%    659.4n ±  0%   +1.91% (p=0.001 n=10)
Coloring/triangle/RecursiveLargestFirst-8                1.171µ ±  0%    1.171µ ±  0%        ~ (p=0.822 n=10)
Coloring/triangle/SanSegundo-8                           1.844µ ±  0%    1.869µ ±  0%   +1.38% (p=0.000 n=10)
Coloring/triangle/WelshPowell-8                          780.7n ±  0%    793.0n ±  0%   +1.58% (p=0.000 n=10)
Coloring/square/Dsatur-8                                 2.226µ ±  2%    2.268µ ±  0%   +1.89% (p=0.003 n=10)
Coloring/square/DsaturExact-8                            10.94µ ±  0%    11.11µ ±  0%   +1.53% (p=0.000 n=10)
Coloring/square/Randomized-8                             867.5n ±  0%    884.6n ±  0%   +1.98% (p=0.000 n=10)
Coloring/square/RecursiveLargestFirst-8                  1.451µ ±  0%    1.448µ ±  0%   -0.24% (p=0.000 n=10)
Coloring/square/SanSegundo-8                             2.843µ ±  0%    2.889µ ±  0%   +1.64% (p=0.000 n=10)
Coloring/square/WelshPowell-8                            1.034µ ±  3%    1.050µ ±  0%   +1.60% (p=0.018 n=10)
Coloring/cycle-5/Dsatur-8                                3.263µ ±  0%    3.302µ ±  0%   +1.18% (p=0.000 n=10)
Coloring/cycle-5/DsaturExact-8                           19.39µ ±  0%    19.58µ ±  1%   +0.95% (p=0.000 n=10)
Coloring/cycle-5/Randomized-8                            1.139µ ±  0%    1.155µ ±  0%   +1.40% (p=0.000 n=10)
Coloring/cycle-5/RecursiveLargestFirst-8                 2.154µ ±  0%    2.148µ ±  1%   -0.26% (p=0.049 n=10)
Coloring/cycle-5/SanSegundo-8                            4.200µ ±  2%    4.240µ ±  0%   +0.93% (p=0.022 n=10)
Coloring/cycle-5/WelshPowell-8                           1.381µ ±  0%    1.392µ ±  0%   +0.80% (p=0.000 n=10)
Coloring/cycle-6/Dsatur-8                                4.355µ ±  0%    4.426µ ±  0%   +1.63% (p=0.000 n=10)
Coloring/cycle-6/DsaturExact-8                           18.53µ ±  0%    18.66µ ±  0%   +0.74% (p=0.000 n=10)
Coloring/cycle-6/Randomized-8                            1.395µ ±  1%    1.409µ ±  0%   +1.04% (p=0.001 n=10)
Coloring/cycle-6/RecursiveLargestFirst-8                 2.529µ ±  0%    2.520µ ±  0%   -0.34% (p=0.005 n=10)
Coloring/cycle-6/SanSegundo-8                            5.634µ ±  0%    5.701µ ±  0%   +1.19% (p=0.000 n=10)
Coloring/cycle-6/WelshPowell-8                           1.686µ ±  0%    1.704µ ±  0%   +1.04% (p=0.000 n=10)
Coloring/wheel-5/Dsatur-8                                3.493µ ±  0%    3.535µ ±  0%   +1.19% (p=0.000 n=10)
Coloring/wheel-5/DsaturExact-8                           23.15µ ±  1%    23.32µ ±  0%   +0.75% (p=0.002 n=10)
Coloring/wheel-5/Randomized-8                            1.244µ ±  0%    1.261µ ±  0%   +1.33% (p=0.000 n=10)
Coloring/wheel-5/RecursiveLargestFirst-8                 2.435µ ±  0%    2.422µ ±  0%   -0.51% (p=0.000 n=10)
Coloring/wheel-5/SanSegundo-8                            5.236µ ±  0%    5.294µ ±  1%   +1.11% (p=0.000 n=10)
Coloring/wheel-5/WelshPowell-8                           1.414µ ±  0%    1.434µ ±  0%   +1.45% (p=0.000 n=10)
Coloring/wheel-6/Dsatur-8                                4.911µ ±  1%    4.966µ ±  0%   +1.12% (p=0.014 n=10)
Coloring/wheel-6/DsaturExact-8                           38.40µ ±  0%    38.61µ ±  0%   +0.54% (p=0.000 n=10)
Coloring/wheel-6/Randomized-8                            1.616µ ±  0%    1.633µ ±  2%   +1.08% (p=0.000 n=10)
Coloring/wheel-6/RecursiveLargestFirst-8                 3.373µ ±  0%    3.357µ ±  0%   -0.50% (p=0.000 n=10)
Coloring/wheel-6/SanSegundo-8                            7.642µ ±  0%    7.716µ ±  0%   +0.97% (p=0.000 n=10)
Coloring/wheel-6/WelshPowell-8                           1.829µ ±  1%    1.850µ ±  0%   +1.18% (p=0.001 n=10)
Coloring/wheel-7/Dsatur-8                                6.403µ ±  0%    6.489µ ±  0%   +1.34% (p=0.000 n=10)
Coloring/wheel-7/DsaturExact-8                           39.86µ ±  0%    40.20µ ±  0%   +0.86% (p=0.000 n=10)
Coloring/wheel-7/Randomized-8                            1.972µ ±  0%    1.993µ ±  0%   +1.04% (p=0.000 n=10)
Coloring/wheel-7/RecursiveLargestFirst-8                 4.010µ ±  1%    3.991µ ±  0%   -0.47% (p=0.000 n=10)
Coloring/wheel-7/SanSegundo-8                            10.30µ ±  0%    10.38µ ±  0%   +0.79% (p=0.000 n=10)
Coloring/wheel-7/WelshPowell-8                           2.204µ ±  0%    2.232µ ±  0%   +1.29% (p=0.000 n=10)
Coloring/sudoku_board/Dsatur-8                           21.39m ±  0%    21.47m ±  0%   +0.34% (p=0.000 n=10)
Coloring/sudoku_board/DsaturExact-8                      85.97m ±  0%    86.19m ±  1%        ~ (p=0.739 n=10)
Coloring/sudoku_board/Randomized-8                       583.0µ ±  2%    580.8µ ±  0%   -0.37% (p=0.000 n=10)
Coloring/sudoku_board/RecursiveLargestFirst-8            3.856m ±  0%    3.901m ±  0%   +1.16% (p=0.000 n=10)
Coloring/sudoku_board/SanSegundo-8                       25.49m ±  0%    25.71m ±  1%   +0.89% (p=0.000 n=10)
Coloring/sudoku_board/WelshPowell-8                      687.4µ ±  0%    689.5µ ±  0%   +0.30% (p=0.000 n=10)
Coloring/sudoku_problem/Dsatur-8                         32.53m ±  2%    32.26m ±  0%   -0.81% (p=0.000 n=10)
Coloring/sudoku_problem/DsaturExact-8                     4.297 ± 94%     5.256 ± 65%        ~ (p=0.529 n=10)
Coloring/sudoku_problem/Randomized-8                     824.9µ ±  0%    821.6µ ±  0%   -0.41% (p=0.000 n=10)
Coloring/sudoku_problem/RecursiveLargestFirst-8          5.718m ±  1%    5.817m ±  0%   +1.74% (p=0.002 n=10)
Coloring/sudoku_problem/SanSegundo-8                     35.45m ±  0%    35.65m ±  1%   +0.57% (p=0.000 n=10)
Coloring/sudoku_problem/WelshPowell-8                    964.7µ ±  0%    961.0µ ±  0%   -0.38% (p=0.000 n=10)
Coloring/cs_shc/Dsatur-8                                 3.383µ ±  0%    3.424µ ±  0%   +1.23% (p=0.000 n=10)
Coloring/cs_shc/DsaturExact-8                            20.63µ ±  1%    20.78µ ±  0%   +0.76% (p=0.022 n=10)
Coloring/cs_shc/Randomized-8                             1.216µ ±  0%    1.230µ ±  0%   +1.19% (p=0.000 n=10)
Coloring/cs_shc/RecursiveLargestFirst-8                  2.429µ ±  0%    2.419µ ±  2%   -0.41% (p=0.022 n=10)
Coloring/cs_shc/SanSegundo-8                             5.180µ ±  0%    5.230µ ±  0%   +0.97% (p=0.000 n=10)
Coloring/cs_shc/WelshPowell-8                            1.435µ ±  0%    1.456µ ±  0%   +1.46% (p=0.000 n=10)
Coloring/gis_hc/Dsatur-8                                 4.029µ ±  1%    4.078µ ±  0%   +1.20% (p=0.001 n=10)
Coloring/gis_hc/DsaturExact-8                            17.03µ ±  0%    17.13µ ±  0%   +0.59% (p=0.000 n=10)
Coloring/gis_hc/Randomized-8                             1.310µ ±  0%    1.325µ ±  1%   +1.15% (p=0.000 n=10)
Coloring/gis_hc/RecursiveLargestFirst-8                  2.353µ ±  0%    2.347µ ±  0%   -0.28% (p=0.000 n=10)
Coloring/gis_hc/SanSegundo-8                             7.548µ ±  0%    7.610µ ±  0%   +0.82% (p=0.000 n=10)
Coloring/gis_hc/WelshPowell-8                            1.673µ ±  1%    1.691µ ±  0%   +1.08% (p=0.011 n=10)
Coloring/gis_shc|rs_shc/Dsatur-8                         2.107µ ±  0%    2.136µ ±  0%   +1.35% (p=0.000 n=10)
Coloring/gis_shc|rs_shc/DsaturExact-8                    9.854µ ±  0%    9.979µ ±  2%   +1.26% (p=0.000 n=10)
Coloring/gis_shc|rs_shc/Randomized-8                     827.9n ±  0%    841.1n ±  0%   +1.60% (p=0.000 n=10)
Coloring/gis_shc|rs_shc/RecursiveLargestFirst-8          1.376µ ±  1%    1.369µ ±  0%   -0.51% (p=0.010 n=10)
Coloring/gis_shc|rs_shc/SanSegundo-8                     2.744µ ±  0%    2.776µ ±  0%   +1.17% (p=0.000 n=10)
Coloring/gis_shc|rs_shc/WelshPowell-8                    1.029µ ±  0%    1.045µ ±  2%   +1.55% (p=0.000 n=10)
Coloring/lf_hc/Dsatur-8                                  6.200µ ±  1%    6.272µ ±  0%   +1.16% (p=0.001 n=10)
Coloring/lf_hc/DsaturExact-8                             32.76µ ±  0%    32.96µ ±  0%   +0.62% (p=0.000 n=10)
…